Move array dereference to top level class
