+++ src/main/java/io/quarkus/gizmo2/creator/BlockCreator.java
@@ -42,6 +42,7 @@ import io.quarkus.gizmo2.desc.ClassMethodDesc;
 import io.quarkus.gizmo2.desc.ConstructorDesc;
 import io.quarkus.gizmo2.desc.FieldDesc;
 import io.quarkus.gizmo2.desc.MethodDesc;
+import io.quarkus.gizmo2.impl.ArrayDeref;
 import io.quarkus.gizmo2.impl.BlockCreatorImpl;
 import io.quarkus.gizmo2.impl.Item;
 import io.quarkus.gizmo2.impl.StaticFieldVarImpl;
@@ -418,7 +419,7 @@ public sealed interface BlockCreator extends SimpleTyped permits BlockCreatorImp
             varHandle = Const.ofFieldVarHandle(fd.desc());
             args = List.of(fd.instance(), expected, update);
             typeDesc = MethodTypeDesc.of(CD_boolean, fd.instance().type(), var.type(), var.type());
-        } else if (var instanceof Item.ArrayDeref ad) {
+        } else if (var instanceof ArrayDeref ad) {
             varHandle = Const.ofArrayVarHandle(ad.type().arrayType());
             args = List.of(ad.array(), ad.index(), expected, update);
             typeDesc = MethodTypeDesc.of(CD_boolean, ad.array().type(), CD_int, var.type(), var.type());
@@ -502,7 +503,7 @@ public sealed interface BlockCreator extends SimpleTyped permits BlockCreatorImp
             varHandle = Const.ofFieldVarHandle(fd.desc());
             args = List.of(fd.instance(), expected, update);
             typeDesc = MethodTypeDesc.of(var.type(), fd.instance().type(), var.type(), var.type());
-        } else if (var instanceof Item.ArrayDeref ad) {
+        } else if (var instanceof ArrayDeref ad) {
             varHandle = Const.ofArrayVarHandle(ad.type().arrayType());
             args = List.of(ad.array(), ad.index(), expected, update);
             typeDesc = MethodTypeDesc.of(var.type(), ad.array().type(), CD_int, var.type(), var.type());
@@ -536,7 +537,7 @@ public sealed interface BlockCreator extends SimpleTyped permits BlockCreatorImp
             varHandle = Const.ofFieldVarHandle(fd.desc());
             args = List.of(fd.instance(), arg);
             typeDesc = MethodTypeDesc.of(var.type(), fd.instance().type(), var.type());
-        } else if (var instanceof Item.ArrayDeref ad) {
+        } else if (var instanceof ArrayDeref ad) {
             varHandle = Const.ofArrayVarHandle(ad.type().arrayType());
             args = List.of(ad.array(), ad.index(), arg);
             typeDesc = MethodTypeDesc.of(var.type(), ad.array().type(), CD_int, var.type());
